Split to serialization and generate phase #50
